requirements.txt.txt
streamlit
pandas
openpyxl
pypdf
python-pptx
langchain-openai
langchain-core